+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `sensitiveOrder` option in `client.executeCommand` and `client.parseCommand`
+- `caseSensitive` option in `client.executeCommand` and `client.parseCommand`
 
 ### Changed